+++ controller/src/main/java/io/pravega/controller/store/stream/tables/serializers/HistoryRecordSerializer.java
@@ -32,13 +32,13 @@ public class HistoryRecordSerializer extends VersionedSerializer.WithBuilder<His
 
     private void read00(RevisionDataInput revisionDataInput, HistoryRecord.HistoryRecordBuilder builder) throws IOException {
         builder.epoch(revisionDataInput.readInt())
-                .segments(revisionDataInput.readCollection(DataInput::readInt, ArrayList::new))
+                .segments(revisionDataInput.readCollection(DataInput::readLong, ArrayList::new))
                 .scaleTime(revisionDataInput.readLong());
     }
 
     private void write00(HistoryRecord history, RevisionDataOutput revisionDataOutput) throws IOException {
         revisionDataOutput.writeInt(history.getEpoch());
-        revisionDataOutput.writeCollection(history.getSegments(), DataOutput::writeInt);
+        revisionDataOutput.writeCollection(history.getSegments(), DataOutput::writeLong);
         revisionDataOutput.writeLong(history.getScaleTime());
     }
